added ability to send json instead of string
