+++ backend/build.gradle
@@ -1,4 +1,6 @@
+
 buildscript {
+
     repositories {
         mavenCentral()
     }
@@ -30,7 +32,6 @@ dependencies {
     compile("org.webjars:stomp-websocket:2.3.3")
     compile("org.webjars:bootstrap:3.3.7")
     compile("org.webjars:jquery:3.1.0")
-
     testCompile("org.springframework.boot:spring-boot-starter-test")
 }
