+++ backend/src/main/java/checksmart/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package checksmart;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
