+++ backend/src/main/java/checksmart/NetworkMessage.java
@@ -1,0 +1,22 @@
+package checksmart;
+
+public class NetworkMessage {
+
+    private String content;
+
+    public NetworkMessage() {
+
+    }
+
+    public NetworkMessage(String content) {
+        this.content = content;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+}
