+++ backend/src/main/java/checksmart/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package hello;
+package checksmart;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -12,7 +12,6 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void configureMessageBroker(MessageBrokerRegistry config) {
-//        config.enableSimpleBroker("/topic");
         config.setApplicationDestinationPrefixes("/app");
     }
