--- backend/src/main/java/hello/Move.java
@@ -1,18 +1,0 @@
-package hello;
-
-public class Move {
-
-    private String content;
-
-    public Move() {
-    }
-
-    public Move(String content) {
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-}
