--- backend/src/main/java/hello/NetworkMessage.java
@@ -1,21 +1,0 @@
-package hello;
-
-public class NetworkMessage {
-
-    private String name;
-
-    public NetworkMessage() {
-    }
-
-    public NetworkMessage(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
